refactor(notes): extract use_peer_merge hook from NoteDetail
